requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary>=2.9.11
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pytz==2023.3
